Improve user experience
Added confirmation on task deletion and task completion to prevent unintentional actions.

+++ client/src/pages/panel/options.js
@@ -7,8 +7,12 @@ const UserOptions = ({task, setToggleModifyTask, setToAuth}) => {
     const removeTask = (event, taskId) => {
         event.preventDefault();
 
-        deleteTask({"_id": taskId})
-        .then(navigate => setToAuth(navigate));
+        const confirm = window.confirm('Delete task?');
+
+        if (confirm) {
+            deleteTask({"_id": taskId})
+            .then(navigate => setToAuth(navigate));
+        }
     };
 
     return (
